refactor: address MastForest cleanup comments
