+++ core/src/mast/node/basic_block_node/arbitrary.rs
@@ -509,7 +509,7 @@ impl Arbitrary for MastForest {
                         }
                     }
 
-                    forest.finish().expect("generated MAST forest should be valid")
+                    forest.build().expect("generated MAST forest should be valid")
                 },
             )
             .boxed()
@@ -605,7 +605,7 @@ impl Arbitrary for Program {
             let node_id = builder.push_node(node_builder).expect("Failed to add node");
             builder.mark_root(node_id);
             let (forest, remapping) =
-                builder.finish_with_id_map().expect("generated program forest should be valid");
+                builder.build_with_id_map().expect("generated program forest should be valid");
             let entrypoint = remapping.get(node_id).expect("entrypoint should be retained");
 
             Program::new(Arc::new(forest), entrypoint)
