+++ crates/package-registry/src/lib.rs
@@ -310,7 +310,7 @@ mod tests {
             .push_node(BasicBlockNodeBuilder::new(vec![Operation::Add]))
             .expect("failed to build basic block");
         builder.mark_root(node_id);
-        let (forest, remapping) = builder.finish_with_id_map().expect("failed to build forest");
+        let (forest, remapping) = builder.build_with_id_map().expect("failed to build forest");
         let node_id = remapping.get(node_id).expect("root node should be retained");
         (forest, node_id)
     }
